chat_app/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.1
openai==1.61.0
SQLAlchemy==2.0.28
Werkzeug==3.0.1
python-dateutil==2.8.2
markupsafe==2.1.5
Jinja2==3.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.7.0
greenlet==3.0.3
typing_extensions==4.12.2
colorama==0.4.6
uuid==1.30
